pyproject.toml
[tool.poetry]
name = "my-unimernet"
version = "0.1.0"
description = ""
authors = ["ParaN3xus <paran3xus007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.5.1"
transformers = "^4.47.1"
ftfy = "^6.3.1"
pillow = "^11.0.0"
albumentations = "^1.4.24"
torchvision = "^0.20.1"
opencv-python = "^4.10.0.84"
omegaconf = "^2.3.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
